move binaries to each runtime crate and make top-level virtual manifest
Current implementation puts runtime binaries into the same crate. It
brings long time to build the whole crate even if one of dependent crate
gets any change. The issue can be resolved when the binaries are put
into each runtime crate.

This commit is for the purpose. As a result, the top-level crate has
virtual manifest for workspace.
