+++ Cargo.toml
@@ -1,31 +1,3 @@
-[package]
-edition = "2018"
-name = "snd-firewire-ctl-services"
-version = "0.1.0"
-# For publishing.
-authors = ["Takashi Sakamoto"]
-description = """
-Service programs to control audio and music unit on IEEE 1394 bus, supported by ALSA
-"""
-license = "GPL-3.0-or-later"
-
-[dependencies]
-libc = "0.2"
-glib = "0.15"
-alsactl = "0.4"
-alsaseq = "0.4"
-hinawa = "0.7"
-hitaki = "0.2"
-core = { path = "runtime/core" }
-digi00x-runtime = { path = "runtime/digi00x" }
-tascam-runtime = { path = "runtime/tascam" }
-fireworks-runtime = { path = "runtime/fireworks" }
-motu-runtime = { path = "runtime/motu" }
-oxfw-runtime = { path = "runtime/oxfw" }
-bebob-runtime = { path = "runtime/bebob" }
-dice-runtime = { path = "runtime/dice" }
-fireface-runtime = { path = "runtime/fireface" }
-
 [workspace]
 members = [
     "runtime/core",
