+++ runtime/fireworks/src/bin/snd-fireworks-ctl-service.rs
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
-// Copyright (c) 2020 Takashi Sakamoto
+// Copyright (c) 2022 Takashi Sakamoto
 
 use {core::cmdline::*, fireworks_runtime::EfwRuntime};
